requirements.txt
# API Backend
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.7.0
pydantic-settings==2.3.0

# Base de données
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pgvector==0.3.2

# Cache
redis[asyncio]==5.0.7

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyotp==2.9.0

# IA Providers
anthropic==0.34.0
openai==1.40.0
httpx==0.27.0

# Data Pipeline
pandas==2.2.2
pyarrow==17.0.0
duckdb==1.0.0

# Utilitaires
python-dotenv==1.0.1
python-multipart==0.0.9
aiofiles==24.1.0
tenacity==9.0.0
apscheduler==3.10.4
bcrypt==4.1.3

# Tests
pytest==8.3.0
pytest-asyncio==0.23.8
pytest-cov==5.0.0
